ANDROID: Fix fsck overflows when handling bad clu_count or vol_length fields.
Bug: 508645975
Test: new exfatprogs test case
Cherrypick-From: https://googleplex-android-review.googlesource.com/q/commit:9befb98273053c48f9b0df810090f4c88e42706a
Merged-In: Ic8223079d3612162b7ddc7c35c53098efef2f707

+++ include/libexfat.h
@@ -24,7 +24,7 @@ typedef __u32 clus_t;
 #define MIN(a, b)	((a) < (b) ? (a) : (b))
 #define MAX(a, b)	((a) > (b) ? (a) : (b))
 
-#define DIV_ROUND_UP(__i, __d)	(((__i) + (__d) - 1) / (__d))
+#define DIV_ROUND_UP(__i, __d)	(((uint64_t)(__i) + (__d) - 1) / (__d))
 
 #define EXFAT_MIN_NUM_SEC_VOL		(2048)
 #define EXFAT_MAX_NUM_SEC_VOL		((2 << 64) - 1)
